improv

+++ core/utils/magisk/uninstall.sh
@@ -1,3 +1,8 @@
-rm -rf /data/adb/service.d/litegapps-post-fs
+if [ -f /data/adb/service.d/litegapps-post-fs ]; then
+	rm -rf /data/adb/service.d/litegapps-post-fs
+fi
+if [ -d /data/adb/litegapps ]; then
+	rm -rf /data/adb/litegapps
+fi
 pm uninstall com.google.android.gms
 pm uninstall com.android.vending
